project3/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-OAuthlib==0.9.6
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0
gunicorn==21.2.0
Werkzeug==2.3.7
